+++ app/Http/Controllers/PurchaseOrderController.php
@@ -5,18 +5,31 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 // call model Purchase Order
 use App\PurchaseOrder;
+//lib db
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Session;
 
 class PurchaseOrderController extends Controller
 {
     public function index()
     {
-    	$PurchaseOrder = PurchaseOrder::all();
-    	return view('purchase_order/index', ['PurchaseOrder' => $PurchaseOrder]);
+		if(Session::get('user_level') == 1){
+			$PurchaseOrder = PurchaseOrder::all();
+    		return view('purchase_order/index', ['PurchaseOrder' => $PurchaseOrder]);
+		}else{
+			return redirect('/dashboard');
+		}
+    	
     }
 
     public function tambah()
     {
-    	return view('purchase_order/tambah');
+		if(Session::get('user_level') == 1){
+			return view('purchase_order/tambah');
+		}else{
+			return redirect('/dashboard');
+		}
+    	
     }
 
     public function tambah_proses(Request $request)
@@ -36,15 +49,25 @@ class PurchaseOrderController extends Controller
 
 	public function hapus($id_purchase_order)
 	{
-	    $purchaseOrder = PurchaseOrder::find($id_purchase_order);
-	    $purchaseOrder->delete();
-	    return redirect('/purchase_order');
+		if(Session::get('user_level') == 1){
+			$purchaseOrder = PurchaseOrder::find($id_purchase_order);
+			$purchaseOrder->delete();
+			return redirect('/purchase_order');
+		}else{
+			return redirect('/dashboard');
+		}
+	    
 	}
 
 	public function edit($id)
 	{
-	   $PurchaseOrder = PurchaseOrder::find($id);
-	   return view('/purchase_order/edit', ['PurchaseOrder' => $PurchaseOrder]);
+		if(Session::get('user_level') == 1){
+			$PurchaseOrder = PurchaseOrder::find($id);
+	   		return view('/purchase_order/edit', ['PurchaseOrder' => $PurchaseOrder]);
+		}else{
+			return redirect('/dashboard');
+		}
+	   
 	}
 
 	public function edit_proses($id, Request $request)
@@ -57,6 +80,8 @@ class PurchaseOrderController extends Controller
 		$PurchaseOrder->customer = $request->customer;
 		$PurchaseOrder->nominal_purchase_order = $request->nominal_purchase_order;
 		$PurchaseOrder->status_delivery = $request->status_delivery;
+		$PurchaseOrder->progress = $request->progress;
+		
 		$PurchaseOrder->save();
 		return redirect('/purchase_order');
 	}
